tests/wpt/tests/css/css-gaps/tentative/parsing/rule-paint-order-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Gaps: gap-rule-paint-order getComputedStyle()</title>
  <link rel="author" title="Javier Contreras Tenorio" href="mailto:javiercon@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-gaps-1/#paint-order">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  <script>
    test_computed_value("gap-rule-paint-order", "row-over-column");
    test_computed_value("gap-rule-paint-order", "column-over-row");
  </script>
</body>
</html>
